Temporarily move netzero in powerwall_control integration
This allows us to bring up the usage of the Netzero API from
the integration without it being in PyPI.

Leave a symlink in netzero/ so that powerwall.py still works.
Do it this way round so that we can continue to copy the
powerwall_control directory across to the Home Assistant
config directory and still expect everything to work.

Eventually netzero.py is expected to move out of this repo,
and into its own repo.