testing: declare a common fixture for mock EnergySite
Change the main code to import netzero in such a way that we
can mock it at a single location.

Now simplify the tests to use mock_energysite.

Remove banned import for `tests`, as we want to import mocks
from confdata.py. We'll figure out how if we need to reorganise
later. For now, it's nice to have the netzero mock setup in one
place.